+++ tests/react/stat-block-weight.test.js
@@ -229,6 +229,25 @@ describe('stat block weight handling', function() {
         expect(screen.getByLabelText("Weight carried").textContent).toEqual("10.25 kg")
     });
 
+    it("adds firearm magazine weight for caseless ammo", async () => {
+        server.use(
+            rest.get('http://localhost/rest/sheets/1/sheetfirearms/', async (req, res, ctx) => {
+                return res(ctx.json(
+                    [factories.firearmFactory({
+                        base: {weight: 5, magazine_weight: 0.75},
+                        scope: null,
+                        ammo: {weight: 12, cartridge_weight: 0.2 },
+                        magazines: [{current: 30, capacity: 40}, {current: 30, capacity: 40}, {current: 40, capacity: 40}]
+                    })]
+                ))
+            }),
+        )
+
+        render(<StatBlock url="/rest/sheets/1/" />)
+        await waitForElementToBeRemoved(() => screen.queryAllByRole("status"), {timeout: 5000})
+        expect(screen.getByLabelText("Weight carried").textContent).toEqual("8.47 kg")
+    });
+
     it("adds scope weight", async () => {
         server.use(
             rest.get('http://localhost/rest/sheets/1/sheetfirearms/', async (req, res, ctx) => {
@@ -311,5 +330,4 @@ describe('stat block weight handling', function() {
         expect(input).toHaveClass("is-valid")
         expect(sheet.getByLabelText("Weight carried").textContent).toEqual("11.00 kg")
     });
-
 });